Add the basis of configuring a cache TTL for course assets.
This PR adds the ability to define a global TTL for unlocked course
assets, which is then sent along with the standard cache controlling
headers to the end user or intermediate proxy/cache.  Locked assets are
not eligible for caching.

This PR also introduces a new piece of middle -- clean_headers -- which
provides a mechanism for scrubbing out particular headers before a
response is returned.  This is required to remove particular headers set
by other middleware which negatively influences the cacheability of
these course assets.